coingecoliker/requirements.txt
bs4
lxml
aiohttp
gunicorn
django-heroku
pymongo
pymongo[srv]
dnspython

requests
tor
stem
fake_useragent
aiohttp_socks
